refactor: extract shared fetchAgentTasks helper
